requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.0
